+++ examples/getting-started/getting-started-project/src/components/modules/Header/index.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Menu, logInfo } from '@hubspot/cms-components';
 import {
   ImageField,
@@ -12,143 +13,7 @@ export function Component({ fieldValues }) {
     <header className={headerStyles.wrapper}>
       <nav>
         <img src={fieldValues.logo.src} alt="Logo" width={100} />
-        {/* <Menu fieldPath="menu" /> */}
-        <div>
-          <span
-            id="hs_cos_wrapper_widget_1717899743787_"
-            className="hs_cos_wrapper hs_cos_wrapper_widget hs_cos_wrapper_type_menu"
-            data-hs-cos-general-type="widget"
-            data-hs-cos-type="menu"
-          >
-            <div
-              id="hs_menu_wrapper_widget_1717899743787_"
-              className="hs-menu-wrapper active-branch flyouts hs-menu-flow-horizontal"
-              role="navigation"
-              data-sitemap-name="default"
-              data-menu-id="53522818668"
-              aria-label="Navigation Menu"
-            >
-              <ul role="menu">
-                <li className="hs-menu-item hs-menu-depth-1" role="none">
-                  <a href="javascript:;" role="menuitem">
-                    Home
-                  </a>
-                </li>
-                <li
-                  className="hs-menu-item hs-menu-depth-1 hs-item-has-children"
-                  role="none"
-                >
-                  <a
-                    href="javascript:;"
-                    aria-haspopup="true"
-                    aria-expanded="false"
-                    role="menuitem"
-                  >
-                    Software
-                  </a>
-                  <ul role="menu" className="hs-menu-children-wrapper">
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        Free HubSpot CRM
-                      </a>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        Marketing Hub
-                      </a>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        Sales Hub
-                      </a>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        Service Hub
-                      </a>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        CMS Hub
-                      </a>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        App Marketplace
-                      </a>
-                    </li>
-                  </ul>
-                </li>
-                <li className="hs-menu-item hs-menu-depth-1" role="none">
-                  <a href="javascript:;" role="menuitem">
-                    Pricing
-                  </a>
-                </li>
-                <li
-                  className="hs-menu-item hs-menu-depth-1 hs-item-has-children"
-                  role="none"
-                >
-                  <a
-                    href="javascript:;"
-                    aria-haspopup="true"
-                    aria-expanded="false"
-                    role="menuitem"
-                  >
-                    Resources
-                  </a>
-                  <ul role="menu" className="hs-menu-children-wrapper">
-                    <li
-                      className="hs-menu-item hs-menu-depth-2 hs-item-has-children"
-                      role="none"
-                    >
-                      <a href="javascript:;" role="menuitem">
-                        Blog
-                      </a>
-                      <ul role="menu" className="hs-menu-children-wrapper">
-                        <li
-                          className="hs-menu-item hs-menu-depth-3"
-                          role="none"
-                        >
-                          <a href="javascript:;" role="menuitem">
-                            Blog Topic Generator
-                          </a>
-                        </li>
-                        <li
-                          className="hs-menu-item hs-menu-depth-3"
-                          role="none"
-                        >
-                          <a href="javascript:;" role="menuitem">
-                            Blog Ideas Generator
-                          </a>
-                        </li>
-                      </ul>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        Customer Stories
-                      </a>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        Developers
-                      </a>
-                    </li>
-                    <li className="hs-menu-item hs-menu-depth-2" role="none">
-                      <a href="javascript:;" role="menuitem">
-                        Ebooks, Guides &amp; More
-                      </a>
-                    </li>
-                  </ul>
-                </li>
-                <li className="hs-menu-item hs-menu-depth-1" role="none">
-                  <a href="javascript:;" role="menuitem">
-                    About
-                  </a>
-                </li>
-              </ul>
-            </div>
-          </span>
-        </div>
+        <Menu fieldPath="menu" />
       </nav>
     </header>
   );
